+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * `chop_equally()` may return fewer than `groups` groups when there are 
   duplicate elements. We now warn when this happens.
+* Bugfix: `chop_n()` could return intervals with fewer than `n` elements 
+  when there were duplicate elements.
 
 # santoku 0.9.1